Add link to API docs
This PR is a proposal for how to close #204.

I actually had no idea where these were, so I thought opting to have them front and center in the Navbar would really help new users explore what Lucky has to offer.

Obviously very open to other thoughts on where this link could/should go, and it doesn't have to only be one place either!

One earmark on this PR is the addition of a "target" attribute for navbar links, so that any links not pointing to luckyframework.org will go to a new tab. It felt kind of easy to be navigated away from the site, which could lead to folks not exploring as thoroughly as they normally would have.