Document default values for model columns
This PR aims to serve as a starting point to close #189.

I'm a little bit fuzzy on how the new functionality in https://github.com/luckyframework/avram/pull/424 actually works in conjunction with the `default` option you can add to a database migration, but I'm hoping this is a good enough starting point to give me a bit of guidance and take some work off of the plate of the folks actually building all of this nifty stuff.