Add current_page? documentation
This PR aims to resolve #371, and encompasses two main changes:

- Rename "Page helpers" to "Formatting helpers", since that's all it contained and it would have been an abrupt switch in content without a larger (and, probably unnecessary) refactor
- Add a new "Finding the current page" section to explain `current_page?`

I might have been a bit heavy-handed with code blocks here, but replicating a short version of the API doc blocks seemed to be the most intuitive way to communicate the nuances of parameter matching.

As always, open to input and suggestions on how to more elegantly explain this.