Fix Markdown Regex Formatting Issue
This PR aims to resolve #434

The issue is likely related to the markdown lib since the last `+` in `/\\w+:\\w+/`.

While we should probably open an issue there, I was able to resolve this and fix a side-effect of the same code by adding line begin/end anchors (`^` and `$`).

This prevents me from using this task as `lucky gen.model MyModel some_column:anotherl33th4ck3r`, which would pass for the non-achored regex.