Add file uploads correction, add deletion
There was an issue where `File.read` returns a `String`, but `Shrine.upload` requires an `IO` input. `File.new` works in an app I built as recently as today from the master branch of the shard.

Additionally, I realized that the code as-is would leave old images around, which isn't much of a problem with disk storage but can become super tricky if you're using AWS like I am.